requirements.txt
alabaster==0.7.13
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
docutils==0.19
fonttools==4.39.4
furo==2023.3.27
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
py==1.11.0
Pygments==2.15.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
PyYAML==6.0
requests==2.29.0
ruff==0.0.263
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==6.2.1
sphinx-basic-ng==1.0.0b1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
threadpoolctl==3.1.0
tomli==2.0.1
urllib3==1.26.15
yapf==0.33.0
zipp==3.15.0
